Added most of the things that were left
Now the game counts the number of objects on the ground, diplays # object collisions with player, sets the materials of game objs when they are hit by player and stops the objs. Also made code better, and removed the terrain Bbox.

Issues still: if you stay in the center spawned bunnies will contact the player for some reason, framerate takes a shit when a lot of objects are drawn on the  screen